Extract types into separate file
